refactor: split ProductList.jsx into smaller reusable components
- Extract ProductControls, ProductTable, ProductRow, Pagination, and optional SummaryCards
- Move related CSS to product.module.css (or separate modules)
- Improve separation of concerns and code readability
- Simplify state management and component responsibilities
- Prepare codebase for better maintainability and future features
