+++ src/components/products/ProductFilters.jsx
@@ -1,0 +1,8 @@
+const ProductFilters = () => {
+    return (
+        <>
+        </>
+    )
+}
+
+export default ProductFilters;
