+++ src/components/ui/ErrorMessage.jsx
@@ -1,0 +1,5 @@
+import common from "../../styles/common.module.css";
+
+const ErrorMessage = ({error}) => <div className={common.error}>Error: {error}</div>
+
+export default ErrorMessage;
