+++ src/components/ui/Loading.jsx
@@ -1,0 +1,5 @@
+import common from "../../styles/common.module.css";
+
+const Loading = () => <div className={common.loadin}>Loading...</div>;
+
+export default Loading;
