+++ src/styles/common.module.css
@@ -79,3 +79,15 @@
     flex-wrap: wrap;
     margin-bottom: 1rem;
   }
+
+  .loading {
+    text-align: center;
+    padding: 1rem;
+    color: #555;
+  }
+
+  .error {
+    color: red;
+    text-align: center;
+    padding: 1rem;
+  }
